requirements.txt
streamlit
pandas
matplotlib
pyyaml
feedparser
transformers
torch
requests
